requirements.txt
tencentcloud-sdk-python>=3.0.0
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0















